move `VRTStack` to `_readers.py`
