+++ tests/test_stack.py
@@ -5,7 +5,7 @@ import numpy.testing as npt
 import pytest
 from osgeo import gdal
 
-from dolphin.stack import VRTStack
+from dolphin._readers import VRTStack
 from dolphin.utils import _get_path_from_gdal_str
 
 # Note: uses the fixtures from conftest.py
