+++ tests/test_workflows_utils.py
@@ -2,12 +2,12 @@ import numpy as np
 import pytest
 from osgeo import gdal
 
-from dolphin import stack
+from dolphin import _readers
 from dolphin.workflows.single import setup_output_folder
 
 
 def test_setup_output_folder(tmpdir, tiled_file_list):
-    vrt_stack = stack.VRTStack(tiled_file_list, outfile=tmpdir / "stack.vrt")
+    vrt_stack = _readers.VRTStack(tiled_file_list, outfile=tmpdir / "stack.vrt")
     out_file_list = setup_output_folder(vrt_stack, driver="GTiff", dtype=np.complex64)
     for out_file in out_file_list:
         assert out_file.exists()
@@ -34,7 +34,7 @@ def test_setup_output_folder(tmpdir, tiled_file_list):
     "strides", [{"x": 1, "y": 1}, {"x": 1, "y": 2}, {"x": 2, "y": 3}, {"x": 4, "y": 2}]
 )
 def test_setup_output_folder_strided(tmpdir, tiled_file_list, strides):
-    vrt_stack = stack.VRTStack(tiled_file_list, outfile=tmpdir / "stack.vrt")
+    vrt_stack = _readers.VRTStack(tiled_file_list, outfile=tmpdir / "stack.vrt")
 
     out_file_list = setup_output_folder(
         vrt_stack, driver="GTiff", dtype=np.complex64, strides=strides
